requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
email-validator
pydantic-settings
psycopg2-binary
pytest
httpx
fastapi-mail